feat: add render modes
