+++ src/engine.mjs
@@ -17,6 +17,7 @@ export { SCENE_W, SCENE_H };
 export const params = {
   fpsCap: 60,
   effect: "gradient",        // "gradient" | "solid" | "fire"
+  renderMode: "duplicate",    // "duplicate" | "extended" | "mirror"
   effects: {},
   post: {
     brightness: 0.8,
@@ -54,7 +55,7 @@ for (const eff of Object.values(effects)) {
 
 export function updateParams(patch){
   for (const [key, value] of Object.entries(patch)) {
-    if (key === "fpsCap" || key === "effect") {
+    if (key === "fpsCap" || key === "effect" || key === "renderMode") {
       params[key] = value;
     } else if (postKeys.has(key)) {
       params.post[key] = value;
@@ -72,6 +73,7 @@ export function updateParams(patch){
 // leftFrame and rightFrame hold RGB float data for each wall
 const leftFrame  = new Float32Array(SCENE_W * SCENE_H * 3);
 const rightFrame = new Float32Array(SCENE_W * SCENE_H * 3);
+const extendedFrame = new Float32Array(SCENE_W * 2 * SCENE_H * 3);
 
 
 // ------- build slices frame -------
@@ -115,10 +117,32 @@ function tick(){
   if (acc >= step) {
     const t = Number(now)/1e9;
     acc = 0;
-
-    // Render scene and duplicate
-    renderScene(leftFrame, t, params);
-    rightFrame.set(leftFrame);
+    const mode = params.renderMode;
+    if (mode === "extended") {
+      const W = SCENE_W * 2;
+      renderScene(extendedFrame, W, SCENE_H, t, params);
+      for (let y = 0; y < SCENE_H; y++) {
+        const src = y * W * 3;
+        const dst = y * SCENE_W * 3;
+        leftFrame.set(extendedFrame.subarray(src, src + SCENE_W * 3), dst);
+        rightFrame.set(extendedFrame.subarray(src + SCENE_W * 3, src + W * 3), dst);
+      }
+    } else {
+      renderScene(leftFrame, SCENE_W, SCENE_H, t, params);
+      if (mode === "mirror") {
+        for (let y = 0; y < SCENE_H; y++) {
+          for (let x = 0; x < SCENE_W; x++) {
+            const src = ((SCENE_H - 1 - y) * SCENE_W + (SCENE_W - 1 - x)) * 3;
+            const dst = (y * SCENE_W + x) * 3;
+            rightFrame[dst] = leftFrame[src];
+            rightFrame[dst + 1] = leftFrame[src + 1];
+            rightFrame[dst + 2] = leftFrame[src + 2];
+          }
+        }
+      } else {
+        rightFrame.set(leftFrame);
+      }
+    }
 
     // Emit SLICES_NDJSON to stdout
     const out = buildSlicesFrame(frame++, cap);
