+++ src/ui/index.html
@@ -60,6 +60,13 @@
         <div id="yaw" class="hslider"><div class="handle"></div></div>
       </label>
       <label>FPS cap <input id="fpsCap" type="range" min="1" max="60" step="1"><span id="fpsCap_v"></span></label>
+      <label>Render mode
+        <select id="renderMode">
+          <option value="duplicate">duplicate</option>
+          <option value="extended">extended</option>
+          <option value="mirror">mirror</option>
+        </select>
+      </label>
     </div>
   </fieldset>
